Move some test out of LibraryCommanderTest to LibraryMembershipCommanderTest
