+++ kifi-backend/modules/shoebox/test/com/keepit/commanders/LibraryCommanderTest.scala
@@ -888,7 +888,6 @@ class LibraryCommanderTest extends TestKitSupport with SpecificationLike with Sh
         implicit val config = inject[PublicIdConfiguration]
 
         val (userIron, userCaptain, userAgent, userHulk, libShield, libMurica, libScience) = setupLibraries
-        val libraryCommander = inject[LibraryCommander]
         val libraryInviteCommander = inject[LibraryInviteCommander]
 
         db.readOnlyMaster { implicit s =>
@@ -972,7 +971,6 @@ class LibraryCommanderTest extends TestKitSupport with SpecificationLike with Sh
           implicit val config = inject[PublicIdConfiguration]
 
           val (userIron, userCaptain, userAgent, userHulk, libShield, libMurica, libScience) = setupInvites
-          val libraryCommander = inject[LibraryCommander]
           val libraryInviteCommander = inject[LibraryInviteCommander]
           val libraryMembershipCommander = inject[LibraryMembershipCommander]
 
@@ -1589,88 +1587,6 @@ class LibraryCommanderTest extends TestKitSupport with SpecificationLike with Sh
           val map = libraryMembershipRepo.userRecentTopFollowedLibrariesAndCounts(Id[User](1), since = DateTime.now().minusDays(1), limit = 2)
           map === Map(Id[Library](1) -> 2, Id[Library](2) -> 1)
         }
-      }
-    }
-
-    "update membership to a library" in {
-      withDb(modules: _*) { implicit injector =>
-        val libraryMembershipCommander = inject[LibraryMembershipCommander]
-        val (user1, user2, user3, user4, lib1) = db.readWrite { implicit s =>
-          val user1 = user().withUsername("nickfury").saved
-          val user2 = user().withUsername("quicksilver").saved
-          val user3 = user().withUsername("scarletwitch").saved
-          val user4 = user().withUsername("somerandomshieldagent").saved
-          val lib1 = library().withOwner(user1).saved // user1 owns lib1
-          membership().withLibraryCollaborator(lib1, user2).saved // user2 is a collaborator lib1 (has read_write access)
-          membership().withLibraryFollower(lib1, user3).saved // user3 is a follower to lib1 (has read_only access)
-
-          libraryMembershipRepo.getWithLibraryIdAndUserId(lib1.id.get, user1.id.get).get.access === LibraryAccess.OWNER
-          libraryMembershipRepo.getWithLibraryIdAndUserId(lib1.id.get, user2.id.get).get.access === LibraryAccess.READ_WRITE
-          (user1, user2, user3, user4, lib1)
-        }
-
-        val userId1 = user1.id.get // owner
-        val userId2 = user2.id.get // collaborator
-        val userId3 = user3.id.get // follower
-        val userId4 = user4.id.get // just a nobody
-
-        // test changing owner access (error)
-        libraryMembershipCommander.updateLibraryMembershipAccess(userId1, lib1.id.get, userId1, None).isRight === false
-
-        // test changing membership that does not exist (error)
-        libraryMembershipCommander.updateLibraryMembershipAccess(userId1, lib1.id.get, userId4, None).isRight === false
-
-        // test changing access to owner (error)
-        libraryMembershipCommander.updateLibraryMembershipAccess(userId1, lib1.id.get, userId2, Some(LibraryAccess.OWNER)).isRight === false
-
-        // test owner demoting access
-        libraryMembershipCommander.updateLibraryMembershipAccess(userId1, lib1.id.get, userId2, Some(LibraryAccess.READ_ONLY)) must beRight
-        db.readOnlyMaster { implicit s =>
-          libraryMembershipRepo.getWithLibraryIdAndUserId(lib1.id.get, userId2).get.access === LibraryAccess.READ_ONLY
-        }
-
-        // test owner promoting access
-        libraryMembershipCommander.updateLibraryMembershipAccess(userId1, lib1.id.get, userId2, Some(LibraryAccess.READ_WRITE)) must beRight
-        db.readOnlyMaster { implicit s =>
-          libraryMembershipRepo.getWithLibraryIdAndUserId(lib1.id.get, userId2).get.access === LibraryAccess.READ_WRITE
-        }
-
-        // test collaborator promoting access
-        libraryMembershipCommander.updateLibraryMembershipAccess(userId2, lib1.id.get, userId3, Some(LibraryAccess.READ_WRITE)) must beRight
-        db.readOnlyMaster { implicit s =>
-          libraryMembershipRepo.getWithLibraryIdAndUserId(lib1.id.get, userId3).get.access === LibraryAccess.READ_WRITE
-        }
-
-        // test collaborator demoting access
-        libraryMembershipCommander.updateLibraryMembershipAccess(userId2, lib1.id.get, userId3, Some(LibraryAccess.READ_ONLY)) must beRight
-        db.readOnlyMaster { implicit s =>
-          libraryMembershipRepo.getWithLibraryIdAndUserId(lib1.id.get, userId3).get.access === LibraryAccess.READ_ONLY
-        }
-
-        // test collaborator promoting access (but library does not allow collabs to invite)
-        db.readWrite { implicit s =>
-          libraryRepo.save(lib1.copy(whoCanInvite = Some(LibraryInvitePermissions.OWNER)))
-        }
-        libraryMembershipCommander.updateLibraryMembershipAccess(userId2, lib1.id.get, userId3, Some(LibraryAccess.READ_WRITE)).isRight === false
-        db.readOnlyMaster { implicit s =>
-          libraryMembershipRepo.getWithLibraryIdAndUserId(lib1.id.get, userId3).get.access === LibraryAccess.READ_ONLY
-        }
-
-        // test collaborator removing access
-        libraryMembershipCommander.updateLibraryMembershipAccess(userId2, lib1.id.get, userId3, None) must beRight
-        db.readOnlyMaster { implicit s =>
-          libraryMembershipRepo.getWithLibraryIdAndUserId(lib1.id.get, userId3) === None
-        }
-
-        // test owner removing access
-        libraryMembershipCommander.updateLibraryMembershipAccess(userId1, lib1.id.get, userId2, None) must beRight
-        db.readOnlyMaster { implicit s =>
-          libraryMembershipRepo.getWithLibraryIdAndUserId(lib1.id.get, userId2) === None
-        }
-
-        // test non-active membership (after removing access) (error)
-        libraryMembershipCommander.updateLibraryMembershipAccess(userId1, lib1.id.get, userId2, None).isRight === false
-
       }
     }
   }
